Don't require duplicating conversation message signals on the viewer.
Since we want GearyController to handle the effects of most user actions,
we've had to duplicate ConversationMessage's signals on
ConversationViewer and chain them down.

Instead, add a message_added & message_removed signal to
ConversationViewer and emit them as appropriate, then have
GearyController connect to those and manually manage hooking/unhooking
itself to ConversationMessage signals as instances are added/removed.

Do this to the link clicked signal to start.

* src/client/application/geary-controller.vala: Add on_message_added and
  on_message_removed handlers, connect them to the main window's
  ConversationViewer instance and added/remove the message's link click
  hander there.

* src/client/conversation-viewer/conversation-viewer.vala: Add
  message_added and message_removed signals, hook them up as messages are
  added and removed. Remove link_selected signal and plumbing.

* src/client/conversation-viewer/conversation-message.vala: Rename
  link_selected to link_activated to match GTK nomenclature better.